backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20
pydantic>=2.0.0
chromadb>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0

